backend/pyproject.toml
[project]
name = "locus-backend"
version = "0.1.0"
description = "Locus - Self-hosted computer vision platform backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "aiosqlite>=0.21.0",
    "python-multipart>=0.0.20",
    "pyjwt>=2.10.0",
    "argon2-cffi>=23.1.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.11",
]
